fix: `browser.storage` should not include unknown keys in the result
As described in the [documentation](https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/storage/StorageArea/get), when the key is not present in the storage, it should be completely missing from the `.get()` result as well. The current mock implementation was returning the key in the result with `undefined` value.

Also the lastly added session storage mock was sharing data with local storage.